Installed tailwindcss/container-queries, tailwind-merge, fluid-tailwind, ts-node; added those tailwind plugins to the tailwind config; renamed order_number columns to be either chapter_display_number or section_display_number or exercise_display_number; changed instances of user_<foo> to instead be student_<foo> when the user could only be a student--user solution to student solution, notes.user_id to notes.student_id, User_Exercise_Progress table to Student_Exercise_Progress, User_Progress table to Student_Progress, various user_id columns to be student_id where applicable, updated all custom indexes to use student instead of user where appropriate; restarted seed.ts file due to increasing complexity and the need for breaking down that complexity into a better organized and manageable setup; created seed_config and various files within it to refactor all of that mentioned complexity; updated index.astro to point to correct current seed file; updated runSeed to point to correct current seed file; replaced old jsdoc style comments in general_utils with actual jsdoc style comments.
