+++ db/seed_config/seed/courses/index.ts
@@ -1,0 +1,7 @@
+import { javascriptFundamentals } from "./javascript-fundamentals";
+
+export const courseData = {
+    courses: [
+        javascriptFundamentals
+    ]
+}
