+++ package.json
@@ -24,6 +24,7 @@
     "@fontsource/lilita-one": "^5.1.0",
     "@fontsource/mononoki": "^5.1.0",
     "@fontsource/righteous": "^5.1.0",
+    "@tailwindcss/container-queries": "^0.1.1",
     "@testing-library/dom": "^10.4.0",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
@@ -32,6 +33,7 @@
     "date-fns": "^4.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "tailwind-merge": "^2.5.4",
     "tailwindcss": "^3.4.14",
     "typescript": "^5.6.3"
   },
@@ -51,6 +53,7 @@
     "@typescript-eslint/parser": "^8.9.0",
     "eslint": "^9.12.0",
     "eslint-config-prettier": "^9.1.0",
+    "fluid-tailwind": "^1.0.3",
     "globals": "^15.11.0",
     "jsdom": "^25.0.1",
     "prettier": "^3.3.3",
