+++ src/pages/index.astro
@@ -2,7 +2,7 @@
 import Layout from '@layouts/Layout.astro'
 import '@utils/temp_file_code'
 import '@utils/general_utils'
-import '@db/seed.ts'
+import '@db/seed_old'
 ---
 
 <Layout title="Login">
