+++ src/scripts/runSeed.ts
@@ -1,4 +1,4 @@
-import seed from '@db/seed'
+import seed from '@db/seed_old'
 
 export default async function () {
   try {
